tests/wpt/web-platform-tests/css/mediaqueries/match-media-parsing.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-syntax-3/#parse-comma-separated-list-of-component-values">
<script type="text/javascript" src="/resources/testharness.js"></script>
<script type="text/javascript" src="/resources/testharnessreport.js"></script>

<script>
function test_parsing(query, expected) {
    if(expected === undefined)
        expected = query;

    test(() => {
        const match = window.matchMedia(query);
        assert_equals(match.media, expected)
    }, "Test parsing '" + query + "' with matchMedia");
}

function test_resolution_parsing() {
    test_parsing("(min-resolution: 1x)");
    test_parsing("(resolution: 2x)");
    test_parsing("(max-resolution: 7x)");

    test_parsing("(resolution: 2dppx)");
    test_parsing("(resolution: 600dpi)");
    test_parsing("(resolution: 77dpcm)");
}

test_parsing("", "");
test_parsing(" ", "");
test_parsing("all", "all");
test_parsing(" all", "all");
test_parsing("   all   ", "all");
test_parsing("all,all", "all, all");
test_parsing(" all , all ", "all, all");
test_parsing("(color)", "(color)");
test_parsing("(color", "(color)");
test_parsing(" (color)", "(color)");
test_parsing(" ( color  )  ", "(color)");
test_parsing(" ( color   ", "(color)");
test_parsing("color)", "not all");
test_parsing("  color)", "not all");
test_parsing("  color ), ( color", "not all, (color)");
test_parsing(" foo ", "foo");
test_parsing(",", "not all, not all");
test_parsing(" , ", "not all, not all");
test_parsing(",,", "not all, not all, not all");
test_parsing("  ,  ,  ", "not all, not all, not all");
test_parsing(" foo,", "foo, not all");

test_resolution_parsing();
</script>
